requirements.txt
tweepy
python-dotenv
openai

